Methods for objects of class "evmiss_sim_study"

Methods for objects of class "evmiss_sim_study" returned from sim_study.

Usage

# S3 method for class 'evmiss_sim_study'
plot(
  x,
  what = c("return", "mu", "sigma", "xi", "all"),
  distn = c("marginal", "joint"),
  approach = c("full", "adjust", "naive", "discard"),
  return_period = 100,
  vertical = TRUE,
  main = c("full", "adjust", "naive", "discard"),
  penultimate = TRUE,
  line_col = c("orange", "purple", "red", "blue"),
  line_lty = c(1, 1, 2, 3),
  legend = 4,
  mar = c(5, 5, 2, 2),
  layout = FALSE,
  no_ylab = FALSE,
  ...
)

# S3 method for class 'evmiss_sim_study'
summary(
  object,
  ...,
  what = c("all", "return"),
  return_period,
  statistics,
  meanSE = FALSE,
  log = FALSE,
  log_return = FALSE,
  vsfull = TRUE
)

# S3 method for class 'summary.evmiss_sim_study'
print(x, quiet = FALSE, ...)

tab(object, ...)

# S3 method for class 'evmiss_sim_study'
tab(object, return_period, ...)

Arguments

x, object

An object inheriting from class "evmiss_sim_study", a result of a call to the function sim_study. For plot.evmiss_sim_study, further arguments to be passed to plot.default and abline, such as graphical parameters.

what

A character scalar. What should be plotted? See distn for further details. The default is "return" for the plot method and "all" for the summary method.

"return": estimates of return levels using the approaches "full", "adjust", "naive" and "discard".

"mu" , "sigma", "xi" or "all". The name of the parameter whose estimates are plotted. Use "all" to select all 3 GEV parameters.

distn

A character scalar. The default is "marginal".

"marginal": 4 histograms of point estimates.

If what = "return" then 4 histograms of the point estimates of the required return level are produced, one plot for each of the approaches "full", "adjust", "naive" and "discard".

If what = "all" then 3 histograms of the respective point estimates of the GEV parameters \(\mu\), \(\sigma\) and \(\xi\) are produced, for the approach supplied in approach.

if what = "mu", "sigma" or "xi" then 4 histograms of the point estimates of the chosen GEV parameter are produced, one for each of the approaches "full", "adjust", "naive" and "discard".

"joint":

If what = "return" or what = "mu", "sigma" or "xi" then pairs is used to produce a matrix of scatter plots to compare the estimates produced by the approaches.

If what = "all" then 3 scatter plots are produced, one for each of the GEV parameters \(\mu\), \(\sigma\) and \(\xi\), to compare the estimates from the 2 approaches supplied in approach.

approach

A character scalar or vector. A proper subset of c("full", "adjust", "naive", "discard"). Only relevant if what is "all". If distn = "marginal" then the default is "full". If distn = "joint" then the default is c("full", "adjust").

return_period

Either a numeric scalar (for plot.evmiss_sim_study) or a numeric vector (for summary.evmiss_sim_study or tab.evmiss_sim_study). A return period or periods, in numbers of blocks, for which to compare the four estimation approaches. All values must be greater than 1.

vertical

logical scalar. Only relevant if distn = "marginal" or if distn = "joint" and what == "all". If vertical = TRUE then the plots are arranged vertically. Otherwise, 4 plots are 2 rows of 2 plots and 3 plots are arranged horizontally.

main

A character vector. Titles for the histograms.

penultimate

A logical scalar. If distn = "marginal" should lines be added to plots to indicate the GEV parameter values, or the implied return levels, from a penultimate approximation based on the simulation distribution and the block length. See mev::smith.penult.

line_col

A vector giving the colours of lines added to plots when distn = "marginal". line_col[1]: sample median of estimates; line_col[2]: sample mean of estimates; line_col[3]: true return level (if what = "return"); line_col[4]: value from an penultimate approximation (if penultimate = TRUE).

line_lty

A numeric vector giving the line types of these vertical lines.

legend

A numeric vector of length 4. In which of the 4 plots should a legend relating to the vertical lines be placed. The default is in the final plot, the "discard" plot.

mar

A numeric vector. The number of lines of margin on the four sides of plots. See par.

layout

A logical scalar. Set layout = TRUE if layout is used to create a multi-plot layout external to the call to plot.evmiss_sim_study.

no_ylab

A logical scalar. IF no_ylab = TRUE then no axis labels are placed on the vertical axis.

...

Further arguments. Only used in the following cases.

plot.evmiss_sim_study: to pass graphical parameters to the graphical functions plot, hist.

print.summary.evmiss_sim_study: if quiet = FALSE, to pass arguments to print, such as digits.

tab.evmiss_sim_study: to pass arguments to summary.evmiss_sim_study, such as vsfull.

statistics

A function used to calculate statistics to summarise the differences between estimates and the quantities described in vsfull below. The object returned from this function should be a named vector. See Examples. If statistics is missing then the default is to calculate the following sample statistics: mean, standard deviation, root mean squared error, median, inter-quartile range, mean absolute error and isNA, the number of simulations for which the GEV parameter MLEs are missing owing to a convergence problem.

meanSE

A logical scalar. If meanSE = TRUE and vsfull = FALSE then the sample mean of the estimated standard errors over all simulated datasets is added to the output summary statistics. Otherwise, this is not added because a comparison of the sample standard deviation of an estimator to meanSE is only valid for vsfull = FALSE.

log

A logical scalar. If log = TRUE then make comparisons for the GEV scale parameter \(\sigma\) based on estimate of \(\log \sigma.\) Otherwise, based them on estimates of \(\sigma\).

log_return

A logical scalar. If log_return = TRUE then make comparisons for return levels on the log scale. Otherwise, use the original return level scale.

vsfull

A logical scalar. If vsfull = TRUE then comparisons are made relative to the estimates from the "full" case, that is, fitting to data with no missing values. If vsfull = FALSE then comparisons are made relative to the true values (what = "return") or penultimate approximations to the GEV parameter values (what = "all").

quiet

A logical scalar. If quiet = TRUE then do not print the summary to the console.

Value

Plot method. No return value, only the plots are produced.

Summary method. A list object, with class c("summary.evmiss_sim_study", "list"), containing the original function call to summary.evmiss_sim_study and the following components, depending on what and vsfull.

If what = "all", matrices full_gev, adjust_gev and naive_gev. Columns 1 to 3 contain the values of the summary statistics relating to the estimates of \(\mu\), \(\sigma\), \(\xi\) respectively.

If what = "return", matrices full_rl, adjust_rl, naive_rl. The columns contain the values of the summary statistics for return levels corresponding to the return periods in return_period.

The object is printed by print.summary.evmiss_sim_study. The rows in the matrices described above are interleaved so that values for the same summary statistic are printed in adjacent rows. If vsfull = TRUE then all the values in full_gev and full_rl are zero and therefore are not included when printing. If the elements of the vector returned from the function statistics are named, then these names are included in the row names. Otherwise, the names used are stat1, stat2 etc.

Details

Plot method. The S3 plot method provides plots to compare these 4 approaches based on the point estimates that they produce for return levels and GEV parameters. Optionally, if penultimate = TRUE, comparisons can be made with a penultimate GEV approximation to the true distribution of the relevant block maxima.

The easiest way to appreciate the various plotting options that are available is to look at the examples in sim_study.

Summary method. A matrix containing the values of the statistics in statistics used to summarise the performance of estimators of GEV parameters (if what = "all") or return levels (if `what = "return"). The columns relate to the GEV parameters or return levels. Each group of 4 rows relates to a different combination of summary statistic and estimator (approach to dealing with missingness), indicated by the row names.

Table method, tab. Creates a table of results in which there is a row for each estimator. If vsfull = FALSE is passed to summary.evmiss_sim_study then the first row relates to the "full" approach. Otherwise, only "adjust", "naive" and "discard" are included.

If return_period is supplied then the statistics concern estimation of these return periods and the statistics included are mean, standard deviation, root mean squared error, median, inter-quartile range and mean absolute error.

Otherwise, the statistics concern the estimation of the GEV parameters and the statistics included are mean, standard deviation and root mean squared error.

Examples

See sim_study.

See also